Major modifications in pipeline
The tree no longer projects the keys: this lies to the scene manager.
What's more, the tree grows using help functions instead of direct access.
This will allow me to use the growing function as an external function to
the class, if ever c++ classes don't end up fitting well with our wrappers.

There are still bugs in the code, and the tree building might change as it
seems there is still an issue with it.